==> 10 Cross_Validation_Practice.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pickle
with open('titanic_step3_feature_onehot_encoding.pickle', 'rb') as pickle_filename:
    df_onehot = pickle.load(pickle_filename)
with open('titanic_step3_feature_encoding_y_pickle', 'rb') as pickle_filename:
    y_train = pickle.load(pickle_filename)

ntrain = 891
X_train, X_test = df_onehot[:ntrain], df_onehot[ntrain:]
X_train.head()

### K-Fold test

from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
from sklearn.neighbors import KNeighborsClassifier
knn = KNeighborsClassifier(n_neighbors=13)
k_fold = KFold(n_splits=10, shuffle=True, random_state=0)
score = cross_val_score(knn, X_train, y_train, cv=k_fold, n_jobs=-1, scoring='accuracy')
score

import numpy as np
np.mean(score) * 100

### Stratified K-Fold test

import pandas as pd
from sklearn.datasets import load_iris
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import KFold

iris_data = load_iris()
dir(iris_data)

iris_df = pd.DataFrame(data=iris_data, columns=iris_data.feature_names)

iris_df['species'] = iris_data.target
iris_df['species'].value_counts()

skfold = StratifiedKFold(n_splits=3)

for train_index, test_index in skfold.split(iris_df, iris_df['species']):
    trainset = iris_df['species'].iloc[train_index]
    testset = iris_df['species'].iloc[test_index]
    print('train set \n', trainset.value_counts())
    print('test set\n', testset.value_counts())


kfold = KFold(n_splits=3)

# KFold.split(X): X 는 훈련 데이터
for train_index, test_index in kfold.split(iris_df):
    trainset = iris_df['species'].iloc[train_index]
    testset = iris_df['species'].iloc[test_index]
    print('trainset\n', trainset.value_counts())
    print('testset\n', testset.value_counts())
